Comment out large body of testing code to verify --use-shmem cases easily.

+++ test.sh
@@ -84,7 +84,7 @@ function record_elapsed_time() {
       # Provide wider test-tag for nightly tests which print verbose descriptions
       fmtstr="%-80s""${fmtstr}"
    else
-      fmtstr="%-50s""${fmtstr}"
+      fmtstr="%-60s""${fmtstr}"
    fi
 
    # Log a line in the /tmp log-file; for future cat of summary output
@@ -650,9 +650,6 @@ fi
 # Run all the unit-tests first, to get basic coverage
 run_with_timing "Fast unit tests" "$BINDIR"/unit_test
 
-# Re-run all the unit-tests first using shared-memory support.
-run_with_timing "Fast unit tests using shared memory" "$BINDIR"/unit_test --use-shmem
-
 # ------------------------------------------------------------------------
 # Run mini-unit-tests that were excluded from bin/unit_test binary:
 # ------------------------------------------------------------------------
@@ -663,12 +660,10 @@ run_with_timing "Fast unit tests using shared memory" "$BINDIR"/unit_test --use-
 
 test_slow_unit_tests ""
 
-test_slow_unit_tests "--use-shmem"
-
 UNIT_TESTS_DB_DEV="unit_tests_db"
-if [ -f ${UNIT_TESTS_DB_DEV} ]; then
-    rm ${UNIT_TESTS_DB_DEV}
-fi
+if [ -f ${UNIT_TESTS_DB_DEV} ]; then rm ${UNIT_TESTS_DB_DEV}; fi
+
+if false; then	# Skip this set of tests ... for now.
 
 key_size=8
 run_with_timing "Functionality test, key size=${key_size} bytes" \
@@ -731,6 +726,21 @@ run_with_timing "Log test" \
 
 run_with_timing "Filter test" \
         "$BINDIR"/driver_test filter_test --seed "$SEED"
+
+fi	#end if false;
+
+# ##############################################################
+# Re-run [some] tests using shared memory in this section.
+# ##############################################################
+
+run_with_timing "Fast unit tests using shared memory" "$BINDIR"/unit_test --use-shmem
+test_slow_unit_tests "--use-shmem"
+if [ -f ${UNIT_TESTS_DB_DEV} ]; then rm ${UNIT_TESTS_DB_DEV}; fi
+
+key_size=8
+run_with_timing "Functionality test, key size=${key_size} bytes, using shared memory" \
+        "$BINDIR"/driver_test splinter_test --functionality 1000000 100 \
+                                            --use-shmem --key-size ${key_size} --seed "$SEED"
 
 record_elapsed_time ${testRunStartSeconds} "All Tests"
 echo ALL PASSED
